Removed warnings when compiling on linux.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2592 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/robots/default_robot.cpp
@@ -522,7 +522,7 @@ void DefaultRobot::handleNitro()
     // (i.e. more than 2) nitro, use it.
     // -------------------------------------------------
     const unsigned int num_karts = race_manager->getNumKarts();
-    if(getPosition()== num_karts && getEnergy()>2.0f)
+    if(getPosition()== (int)num_karts && getEnergy()>2.0f)
     {
         m_controls.wheelie = true;
         return;
